Code refactor of Json converters.
