+++ GoogleCloudExtension/GoogleCloudExtension/GoogleCloudExtension.csproj
@@ -246,10 +246,9 @@
       <DependentUpon>TimeBox.xaml</DependentUpon>
     </Compile>
     <Compile Include="StackdriverLogsViewer\TimeBoxViewModel.cs" />
-    <Compile Include="StackdriverLogsViewer\TreeViewConverters\ComplexPropertyMethodToValueConverter.cs" />
-    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JArrayLengthConverter.cs" />
-    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JPropertyDataTemplateSelector.cs" />
-    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JValueConverter.cs" />
+    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JsonComplexMethodConverter.cs" />
+    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JsonDataTemplateSelector.cs" />
+    <Compile Include="StackdriverLogsViewer\TreeViewConverters\JValueObjectConverter.cs" />
     <Compile Include="StackdriverLogsViewer\TreeViewConverters\LogItemConverter.cs" />
     <Compile Include="StackdriverLogsViewer\TreeViewConverters\MethodToValueConverter.cs" />
     <Compile Include="StackdriverLogsViewer\TreeViewConverters\TreeObjectNode.cs" />
